# api/openapi.yaml
openapi: 3.1.0
info:
  title: jtpost API
  version: 0.10.0
  description: |
    REST API для jtpost — CLI/web-сервиса управления постами для Telegram.

    Текущая стабильная версия — `/api/v1/`. Legacy `/api/` сохранён для backward-compat.

    Authentication:
    - `bearerAuth` — Personal Access Token (`jtpat_<8>_<24>`) в `Authorization: Bearer ...`.
    - `cookieAuth` — Session cookie `jtpost_session` (HttpOnly, Secure, SameSite=Lax).
  license:
    name: MIT
servers:
  - url: /api/v1
    description: Versioned API (stable)
  - url: /api
    description: Legacy (без версии, backward-compat)
...
paths:
  /posts:
    get:
      summary: List posts
      operationId: listPosts
      tags: [posts]
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: status
          in: query
          description: Фильтр по PostStatus (можно повторять)
          schema:
            type: string
        - name: tag
          in: query
          description: Фильтр по тегу (можно повторять)
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [created_at, updated_at, deadline, scheduled_at, title, status]
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 0
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: Список постов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      summary: Create post
      operationId: createPost
      tags: [posts]
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '201':
          description: Post created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /posts/{id}:
    parameters:
      - $ref: '#/components/parameters/PostIdParam'
    get:
      summary: Get post by ID
      operationId: getPost
      tags: [posts]
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        '200':
          description: Post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update post
      operationId: updatePost
      tags: [posts]
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePostRequest'
      responses:
        '200':
          description: Updated post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete post
      operationId: deletePost
      tags: [posts]
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        '204':
          description: Post deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /posts/{id}/publish:
    parameters:
      - $ref: '#/components/parameters/PostIdParam'
    post:
      summary: Publish post
      operationId: publishPost
      tags: [posts]
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        '200':
          description: Published post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /outbox:
    get:
      summary: List outbox entries
      operationId: listOutbox
      tags: [outbox]
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_flight, done, failed]
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
      responses:
        '200':
          description: Outbox entries
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/OutboxFullEntry'}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          description: Outbox not configured
          content:
            text/plain:
              schema: {type: string}
...
  /posts/{id}/queue:
    parameters:
      - $ref: '#/components/parameters/PostIdParam'
    post:
      summary: Enqueue post for async publication
      description: Кладёт пост в outbox-очередь. Worker обработает запись асинхронно.
      operationId: queuePost
      tags: [posts]
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        '202':
          description: Outbox entry created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutboxEntry'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          description: Outbox not configured
          content:
            text/plain:
              schema: {type: string}
  /stats:
    get:
      summary: Get statistics
      operationId: getStats
      tags: [stats]
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        '200':
          description: Stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stats'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /plan:
    get:
      summary: Get publish plan
      operationId: getPlan
      tags: [stats]
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            minimum: 1
            default: 30
      responses:
        '200':
          description: Plan
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlanItem'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /tags:
    get:
      summary: List unique tags
      operationId: listTags
      tags: [stats]
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        '200':
          description: Tags
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /auth/login:
    post:
      summary: Login with email + password
      operationId: authLogin
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Logged in (session cookie set)
          headers:
            Set-Cookie:
              schema:
                type: string
                example: jtpost_session=jts_...; HttpOnly; Secure; SameSite=Lax; Path=/
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /auth/csrf:
    post:
      summary: Refresh CSRF token
      operationId: refreshCsrf
      tags: [auth]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: New CSRF token
          headers:
            X-CSRF-Token:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CsrfResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: PAT
      description: Personal Access Token формата `jtpat_<8>_<24>`
    cookieAuth:
      type: apiKey
      in: cookie
      name: jtpost_session
  parameters:
    PostIdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
        x-go-type: uuid.UUID
        x-go-type-import:
          path: github.com/google/uuid
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    Post:
      type: object
      required:
        - id
        - tenant_id
        - author_id
        - title
        - slug
        - status
        - created_at
        - updated_at
        - revision
        - content
        - external
      properties:
        id:
          type: string
          format: uuid
          x-go-type: uuid.UUID
          x-go-type-import:
            path: github.com/google/uuid
        tenant_id:
          type: string
          format: uuid
          x-go-type: uuid.UUID
          x-go-type-import:
            path: github.com/google/uuid
        author_id:
          type: string
          format: uuid
          x-go-type: uuid.UUID
          x-go-type-import:
            path: github.com/google/uuid
        title:
          type: string
        slug:
          type: string
        status:
          $ref: '#/components/schemas/PostStatus'
        tags:
          type: array
          items:
            type: string
        deadline:
          type: string
          format: date-time
          nullable: true
        scheduled_at:
          type: string
          format: date-time
          nullable: true
        published_at:
          type: string
          format: date-time
          nullable: true
        excerpt:
          type: string
          nullable: true
        cover_image:
          allOf:
            - $ref: '#/components/schemas/Attachment'
          nullable: true
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
        publish_history:
          type: array
          items:
            $ref: '#/components/schemas/PublishAttempt'
        revision:
          type: integer
        revision_sha:
          type: string
          nullable: true
        content:
          type: string
        external:
          $ref: '#/components/schemas/ExternalLinks'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...